change money type from double to BigDecimal

+++ src/main/java/Transaction/Deposit.java
@@ -1,8 +1,10 @@
 package Transaction;
 
+import java.math.BigDecimal;
+
 public class Deposit extends Transaction {
 
-	public Deposit(String performingPersonId, long performingAccountId, double amount) {
+	public Deposit(String performingPersonId, long performingAccountId, BigDecimal amount) {
 		super(TransactionType.DEPOSIT, performingPersonId, performingAccountId, amount);
 		comment = generateComment();
 	}
